fixed skybrowser type and types.ts files
