+++ src/panels/SkyBrowserPanel/WorldWideTelescope/WwtProvider/hooks.ts
@@ -1,6 +1,6 @@
 import { useCallback, useContext, useEffect, useRef, useState } from 'react';
 
-import { Messages } from '@/types/skybrowsertypes';
+import { Messages } from '../../types';
 
 import { WwtContext } from './WwtContext';
